+++ vedo/mesh.py
@@ -925,7 +925,7 @@ class Mesh(Points):
         Stretch mesh between points `q1` and `q2`.
 
         Examples:
-            - [aspring.py](https://github.com/marcomusy/vedo/tree/master/examples/simulations/aspring.py)
+            - [aspring1.py](https://github.com/marcomusy/vedo/tree/master/examples/simulations/aspring1.py)
 
             ![](https://vedo.embl.es/images/simulations/50738955-7e891800-11d9-11e9-85cd-02bd4f3f13ea.gif)